Renomeia MainActivity.kt e altera sua package
