+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.TodoMovies">
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.activity.DetalhesFilmeActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
