--- app/src/main/java/me/vitornascimento/todomovies/MainActivity.kt
@@ -1,20 +1,0 @@
-package me.vitornascimento.todomovies
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import me.vitornascimento.todomovies.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-    private val view by lazy {
-        binding.root
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(view)
-    }
-
-}
